Fixed percent_pos being parsed as an int.
percent_pos property should be a float.

+++ mpv.py
@@ -509,7 +509,7 @@ for name, proptype, access in (
 		('avsync',						float,	'r'),
 		('total-avsync-change',			float,	'r'),
 		('drop-frame-count',			int,	'r'),
-		('percent-pos',					int,	'rw'),
+		('percent-pos',					float,	'rw'),
 		('ratio-pos',					float,	'rw'),
 		('time-pos',					float,	'rw'),
 		('time-start',					float,	'r'),
